Updated writers.

+++ src/main/java/fr/utbm/rngames/keyboard/KeyboardWriter.java
@@ -5,7 +5,6 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
-import java.io.Writer;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.logging.Logger;
@@ -22,17 +21,22 @@ public class KeyboardWriter extends KeyboardListener {
 
 	private final Logger logger = Logger.getLogger(KeyboardWriter.class.getName());
 
-	private final Writer writer;
+	private final URL fileLocation;
+	private final BufferedWriter writer;
 
 	public KeyboardWriter(URL fileLocation) throws IOException {
-		File file = new File(fileLocation.getPath());
-		file.createNewFile();
-		file.deleteOnExit();
+		this.fileLocation = fileLocation;
+		File file = new File(this.fileLocation.getPath());
+
+		if (file.createNewFile()) {
+			file.deleteOnExit();
+		}
 
 		this.writer = new BufferedWriter(new OutputStreamWriter(
 				new FileOutputStream(file), StandardCharsets.UTF_8));
 
-		this.writer.write(Locale.getString(KeyboardWriter.class, "keyboard.file.header") + '\n'); //$NON-NLS-1$
+		this.writer.write(Locale.getString(KeyboardWriter.class, "keyboard.file.header"));
+		this.writer.newLine();
 	}
 
 	@Override
@@ -48,6 +52,7 @@ public class KeyboardWriter extends KeyboardListener {
 	public void nativeKeyPressed(NativeKeyEvent evt) {
 		try {
 			this.writer.write(generateFileEntry("Key Down", evt)); //$NON-NLS-1$
+			this.writer.newLine();
 		} catch (IOException exception) {
 			this.logger.severe(exception.getMessage());
 		}
@@ -57,6 +62,7 @@ public class KeyboardWriter extends KeyboardListener {
 	public void nativeKeyReleased(NativeKeyEvent evt) {
 		try {
 			this.writer.write(generateFileEntry("Key Up", evt)); //$NON-NLS-1$
+			this.writer.newLine();
 		} catch (IOException exception) {
 			this.logger.severe(exception.getMessage());
 		}
@@ -67,9 +73,13 @@ public class KeyboardWriter extends KeyboardListener {
 		// Unused
 	}
 
+	public URL getFileLocation() {
+		return this.fileLocation;
+	}
+
 	private static String generateFileEntry(String eventType, NativeKeyEvent event) {
 		return eventType + CSV_SEPARATOR
 				+ NativeKeyEvent.getKeyText(event.getKeyCode()) + CSV_SEPARATOR
-				+ event.getWhen() + "\n"; //$NON-NLS-1$
+				+ event.getWhen();
 	}
 }
