+++ src/main/java/fr/utbm/rngames/mouse/MouseWriter.java
@@ -9,7 +9,6 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
-import java.io.Writer;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.logging.Logger;
@@ -23,17 +22,22 @@ public class MouseWriter extends MouseListener {
 
 	private final Logger logger = Logger.getLogger(MouseWriter.class.getName());
 
-	private final Writer writer;
+	private final URL fileLocation;
+	private final BufferedWriter writer;
 
 	public MouseWriter(URL fileLocation) throws IOException {
-		File file = new File(fileLocation.getPath());
-		file.createNewFile();
-		file.deleteOnExit();
+		this.fileLocation = fileLocation;
+		File file = new File(getFileLocation().getPath());
+
+		if (file.createNewFile()) {
+			file.deleteOnExit();
+		}
 
 		this.writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(file),
 				StandardCharsets.UTF_8));
 
-		this.writer.write(Locale.getString(MouseWriter.class, "mouse.file.header") + '\n'); //$NON-NLS-1$
+		this.writer.write(Locale.getString(MouseWriter.class, "mouse.file.header"));
+		this.writer.newLine();
 	}
 
 	@Override
@@ -53,7 +57,8 @@ public class MouseWriter extends MouseListener {
 	@Override
 	public void nativeMousePressed(NativeMouseEvent evt) {
 		try {
-			this.writer.write(generateMouseButtonEntry("down", evt)); //$NON-NLS-1$
+			this.writer.write(generateMouseButtonEntry(Locale.getString("mouse.event.pressed"), evt));
+			this.writer.newLine();
 		} catch (IOException exception) {
 			this.logger.severe(exception.getMessage());
 			// System.exit(-1);
@@ -63,7 +68,8 @@ public class MouseWriter extends MouseListener {
 	@Override
 	public void nativeMouseReleased(NativeMouseEvent evt) {
 		try {
-			this.writer.write(generateMouseButtonEntry("up", evt)); //$NON-NLS-1$
+			this.writer.write(generateMouseButtonEntry(Locale.getString("mouse.event.released"), evt));
+			this.writer.newLine();
 		} catch (IOException exception) {
 			this.logger.severe(exception.getMessage());
 			// System.exit(-1);
@@ -74,6 +80,7 @@ public class MouseWriter extends MouseListener {
 	public void nativeMouseMoved(NativeMouseEvent evt) {
 		try {
 			this.writer.write(generateMouseMoveEntry(evt));
+			this.writer.newLine();
 		} catch (IOException exception) {
 			this.logger.severe(exception.getMessage());
 			// System.exit(-1);
@@ -89,18 +96,22 @@ public class MouseWriter extends MouseListener {
 	public void nativeMouseWheelMoved(NativeMouseWheelEvent evt) {
 		try {
 			this.writer.write(generateMouseButtonEntry(evt));
+			this.writer.newLine();
 		} catch (IOException exception) {
 			this.logger.severe(exception.getMessage());
 			// System.exit(-1);
 		}
 	}
 
+	public final URL getFileLocation() {
+		return this.fileLocation;
+	}
+
 	private static String generateMouseButtonEntry(String eventType, NativeMouseEvent evt) {
 		return "mouse " + evt.getButton() + " " +  eventType + CSV_SEPARATOR //$NON-NLS-1$ //$NON-NLS-2$
 				+ evt.getX() + CSV_SEPARATOR
 				+ evt.getY() + CSV_SEPARATOR
-				+ evt.getWhen()
-				+ "\n"; //$NON-NLS-1$
+				+ evt.getWhen();
 	}
 
 	private static String generateMouseButtonEntry(NativeMouseWheelEvent evt) {
@@ -108,22 +119,19 @@ public class MouseWriter extends MouseListener {
 			return "mouse scroll down" + CSV_SEPARATOR //$NON-NLS-1$
 					+ evt.getX() + CSV_SEPARATOR
 					+ evt.getY() + CSV_SEPARATOR
-					+ evt.getWhen()
-					+ "\n"; //$NON-NLS-1$
+					+ evt.getWhen();
 		}
 
 		return "mouse scroll up" + CSV_SEPARATOR //$NON-NLS-1$
 				+ evt.getX() + CSV_SEPARATOR
 				+ evt.getY() + CSV_SEPARATOR
-				+ evt.getWhen()
-				+ "\n"; //$NON-NLS-1$
+				+ evt.getWhen();
 	}
 
 	private static String generateMouseMoveEntry(NativeMouseEvent evt) {
 		return "mouse move" + CSV_SEPARATOR //$NON-NLS-1$
 				+ evt.getX() + CSV_SEPARATOR
 				+ evt.getY() + CSV_SEPARATOR
-				+ evt.getWhen()
-				+ "\n"; //$NON-NLS-1$
+				+ evt.getWhen();
 	}
 }
